move AssertionHandler to dedicated file, remove TestName for now
